Health
Changed the name of the enemymovement component to normalenemymovement component to reflect the way the enemy acts. This will be usefull when adding different behaviour and thus different movement components. I have decided to make the movement component the main component for the AI and their attacks. The amount of health isnt properly displayed yet though. This needs fixing still.