Melee, to be continued
added the possibility to cast three raycasts simultaniously, the idae is to cast two diagnal raycasts and one horizontal raycast to create a cone like attack. still needs improvement though.
There is an issue though, the raycast does not detect the enemy that is patrolling the level. this needs a fix.